New Dataset and cleaning
Added the new dataset and have changed the way certain collums so that we can use it with our model here this being RF where we can't really use strings so we have to change them to numerical values. This is done by using change all references in the dataset. I changed the hereditary_diseases and the job_title. I would also need to do the cities but these are 91 different ones so it can be a bit annoying and long. i'll do that later but this is so d-far what i did. :D